Use socket2 crate
socket2 is the current best way to do advanced socket config without
unsafe code. This lets us remove a lot of our own code using unsafe libc
directly, in udp_unix.rs.

One related change is our UDPSocket's sendto() now returns a Result.
This requires some handling of Result in calling code that is a little
bit wordy -- to emit a diagnostic but not abort, to match the old behavior
(for now).