Clarify that CAP_NET_ADMIN is required
From the current wording of the README, I thought boringtun could run without `CAP_NET_ADMIN`- but it can't at all, failing with:

```
2023-01-04T23:37:51.496375Z ERROR boringtun_cli: Failed to initialize tunnel, error: IOCtl("Operation not permitted")
```

It looks like this originates from `ioctl(fd, TUNSETIFF as _, &ifr)`, which is a privileged operation. 

I think I'm not the only one to make this mistake (https://github.com/cloudflare/boringtun/issues/117), so this clarifies that `CAP_NET_ADMIN` is required.