Upgrade to daemonize 0.5.0 to get rid of boxfnonce
boxfnonce crate is obsolete since Rust 1.35 and cargo audit complains
about it. daemonize 0.5.0 doesn't use it any longer but is incompatible
with earlier version in that the exit_action() modifier is no longer
supported.

Despite the lack of any hint as to how to achieve the same thing with
the newer version, it seems it can be done using daemonize.execute()
instead of damonize.start(), since it will allow the parent to continue
execution past the fork point.

Signed-off-by: David Verbeiren <david.verbeiren@tessares.net>