Reduce locker induced congetion
After migration from v0.5.2 to v0.6.0, a global mutex was introduced
around Peer struct. This lest to major regresion of speed then comparing
these versions on multiple threads.

To fix this, this big lock is now split into several smaller ones,
this is how it was done in v0.5.2, but tunnel is locked with a Mutex.
It was possible to move most palces that rellied on .read() to
immutable state (primarly public_key data, what is built on creation
and never changed).
And as Mutex is way more lean, this leads to major performance gains.

In testing speed is improved by around ~20%.