fix: remove panics from packet processing path (#26)
At present, there are several `panic!`s lurking in the processing path
when the destination buffer for a particular packet isn't big enough.
Whilst in theory this shouldn't happen, panicking and bringing the
entire application down is not very good either. To give the application
on top a better chance of recovering from this error, we replace the
panic with an error.