chore: add warning logs when destination buffer is too small (#27)
Instead of just returning an error, we also add `tracing::warn` logs to
describe the specific mismatch in buffer length. Normally, returning and
logging an error is considered an anti-pattern. But, adding fields to
the `DestinationBufferTooSmall` variant would be a breaking change and
we want to stay semver-compatible for now. See
https://github.com/firezone/boringtun?tab=readme-ov-file#how-to-use-this-fork.