feat: implement handshake jitter (#46)
As per the WireGuard spec
(https://www.wireguard.com/papers/wireguard.pdf#page=14), every time an
expiring timer causes a handshake initiation, a random jitter between 0
and 333ms should be added. This prevents nodes from attempting to
re-initiate a new session at the same time. In Firezone's
strongly-deterministic test suite, this is triggered very easily because
we advance time at the exact same pace.

In order to make use of the deterministic functions added in #45 within
Firezone, this needs to be fixed.

Generating the jitter requires randomness. In order to be deterministic,
a new `rng_seed` parameter is added to the `Tunn::new_at` function. An
observer cannot guess this seed and thus, will be unable to proactively
attempt to initiate handshakes at the same time as the current node.

Fixes: #19.