fix: use correct boolean comparison for scheduled handshake (#48)
In #46, we implemented the handshake jitter required by the WireGuard
spec. Unfortunately, there was a boolean comparison bug where we
compared the deadline, _when_ we need to send the next handshake
incorrectly with the current time such that it was always more or less
sent immediately IF `update_timers_at` was called within the deadline.