fix: prioritize scheduled handshake in `Tunn::update_timers_at` (#50)
In case we have a handshake scheduled and the deadline has surpassed the
current time, sending the handshake is the first thing we want to do.
This fixes a "bug" where we would see duplicate logs for scheduled
handshakes because the evaluation that we need to send a handshake would
run again when `update_timers_at` is called to actually send the
handshake.