fix: ensure responder does not initiate new handshake (#53)
As part of implementing the handshake jitter in #46, we also added a
functionality to initiate a new handshake if the previous session has
already expired. What we overlooked is that only the initiator of the
previous session should do that. Otherwise, both nodes will start to
establish a session and we end up having two active sessions at the same
time.