fix: avoid repeated printing of time-related actions (#55)
Within the `update_timers_at` function, `boringtun` determines
time-related actions such as keep-alives and (scheduled) handshakes.
With the addition of handshake jitter in #46, a handshake may not be
immediately sent after `update_timers_at` is called.

Instead, a handshake is scheduled and returned at a later time (at most
333ms later). In case `update_timers_at` is called again within the
period, the checks would evaluate to the same result, for example, that
a new handshake needs to be sent due to a rekey-timeout. This leads to
confusing logs such as:

```
2025-01-13T06:54:35.561Z DEBUG boringtun::noise::timers: Scheduling new handshake jitter=262.988289ms cid=b3d34a15-55ab-40df-994b-a838e75d65d7
2025-01-13T06:54:35.562Z DEBUG boringtun::noise::timers: HANDSHAKE(REKEY_TIMEOUT) cid=b3d34a15-55ab-40df-994b-a838e75d65d7
2025-01-13T06:54:35.657Z DEBUG boringtun::noise::timers: HANDSHAKE(REKEY_TIMEOUT) cid=b3d34a15-55ab-40df-994b-a838e75d65d7
2025-01-13T06:54:35.657Z DEBUG boringtun::noise::timers: HANDSHAKE(REKEY_TIMEOUT) cid=b3d34a15-55ab-40df-994b-a838e75d65d7
2025-01-13T06:54:35.739Z DEBUG boringtun::noise::timers: HANDSHAKE(REKEY_TIMEOUT) cid=b3d34a15-55ab-40df-994b-a838e75d65d7
2025-01-13T06:54:35.739Z DEBUG boringtun::noise::timers: HANDSHAKE(REKEY_TIMEOUT) cid=b3d34a15-55ab-40df-994b-a838e75d65d7
```

To fix this, we early-exit from `update_timers_at` with
`TunnResult::Done` in case a handshake is scheduled but the deadline is
not reached yet. The delay here is at most 333ms, therefore, not
performing the other time-related actions for that duration is not
critical as everything else in WireGuard has at most a granularity of
1s.