refactor: express timers as `Instant`s (#58)
At present, all timers within `boringtun` are expressed as `Duration`s
from the point in time when `boringtun` got initialised. Upon every
time-related update, we then have to compute, how much time has passed
since this "init" time.

We can avoid all of this arithmetic if we use `Instant`s directly. These
can be compared directly to the current time (or a point in time in the
future).

Related: #54.