refactor: use `now` everywhere instead of `TimerName::TimeCurrent` (#59)
At present, `boringtun`'s time handling works by storing the "current"
time on every update to `update_timers_at`. This time is then used to
timestamp sessions, keep-alives etc.

This isn't as precise as we can be! With #45, we now track time as an
explicit dependency to all functions, meaning functions such as handling
an incoming packet already have access to precisely the current time! We
can leverage this to use that provided timestamp instead of keeping
around the `TimeCurrent` timer.

This results in a few more parameters here and there but makes reasoning
about time within `boringtun` easier.

Depends-On: #58.