refactor: track session timer in `Session` (#60)
The time when a session was created is currently tracked in an array in
`Timers` whereas the actual `Sessions` are tracked in an array in
`Tunn`. The only way that these are correlated is by using the same
index into the array. An expired session is removed from the latter
array but its timer is simply reset to the current time because Rust
doesn't allow a value in an array to be uninitialised.

To fix this, we remove the 2nd array and move the timestamp when a
session got initialised into the `Session` itself. By co-locating this,
this timestamp is automatically gone when the `Session` gets
de-allocated, preventing lingering timer state after a session has
expired.

Depends-On: #59.