fix: don't send packets on almost expired sessions (#68)
WireGuard has several internal timers. For example, if we have
previously sent a data packet on a session and the session becomes older
than 120s, we automatically perform a re-key. This ensures session keys
are constantly rolled over on sessions that are actively used. If a
session is not used at all, it expires after 180s.

There is a race condition where a packet may be sent on a previously
entirely idle session but just before it expires, i.e. at second 179 of
the session. In that case, the sending party will still happily encrypt
it but depending on the latency to the receiver, the session may be
expired on their end by the time it arrives.

To fix this, we introduce a new timer called `SHOULD_NOT_USE_AFTER_TIME`
which is defined as `REJECT_AFTER_TIME` - `KEEPALIVE_TIMEOUT`. If we
attempt to send a packet on a session that is about to expire and we are
the initiator of the current session, we instead buffer the packet and
emit a new handshake.

It is important that only the current initiator performs this check,
otherwise we would introduce another race condition where the responder
would initiate the next session in case it wants to sends an
application-level reply just after `SHOULD_NOT_USE_AFTER_TIME`.