refactor: directly compute when to make new handshake (#69)
When we send a data packet and don't receive any replies for a certain
amount of time, we want to initiate a new handshake. This is currently
implemented by remembering the time we sent the first unreplied packet.
That timer is cleared every time we receive a packet.

The functionality we want to implement is easier to understand if we
directly compute the time when the new handshake is due. This allows us
to avoid recomputing that timestamp later on and we can just check if
`now` has surpassed this timestamp.