fix: reference released version of crypto dependencies (#81)
Referencing pre-releases should ALWAYS happen by pinning them (with
`=`). This is because from a semver PoV, the pre-release component has
no semantic meaning and is only compared lexically:
https://semver.org/#spec-item-11

Pre-releases allow for breaking changes and thus, a breaking change can
sneak into a dependency graph very easily if the pre-release version is
not pinned.

The referenced crates have already released the corresponding version so
we can just switch to that.